tools/chronicle_sim_v2/requirements.txt
PySide6>=6.6.0
pydantic>=2.5.0
pydantic-ai>=0.0.20
PyYAML>=6.0.1
httpx>=0.27.0
openai>=1.0.0
chromadb>=0.4.24
numpy>=1.24.0
json-repair>=0.58
